[build] Formalise mechanism for accessing absolute symbols
In a position-dependent executable, where all addresses are fixed
at link time, we can use the standard technique as documented by
GNU ld to get the value of an absolute symbol, e.g.:

    extern char _my_symbol[];

    printf ( "Absolute symbol value is %x\n", ( ( int ) _my_symbol ) );

This technique may not work in a position-independent executable.
When dynamic relocations are applied, the runtime addresses will no
longer be equal to the link-time addresses.  If the code to obtain the
address of _my_symbol uses PC-relative addressing, then it will
calculate the runtime "address" of the absolute symbol, which will no
longer be equal the the link-time "address" (i.e. the correct value)
of the absolute symbol.

Define macros ABS_SYMBOL(), ABS_VALUE_INIT(), and ABS_VALUE() that
provide access to the correct values of absolute symbols even in
position-independent code, and use these macros wherever absolute
symbols are accessed.

+++ src/arch/x86/transitions/librm_mgmt.c
@@ -426,7 +426,7 @@ void setup_sipi ( unsigned int vector, uint32_t handler,
 	memcpy ( &sipi_regs, regs, sizeof ( sipi_regs ) );
 
 	/* Copy real-mode handler */
-	copy_to_real ( ( vector << 8 ), 0, sipi, ( ( size_t ) sipi_len ) );
+	copy_to_real ( ( vector << 8 ), 0, sipi, sipi_len );
 }
 
 PROVIDE_IOMAP_INLINE ( pages, io_to_bus );
